backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.3.0
firebase-admin==6.2.0
numpy==1.24.3
pandas==2.0.3
pillow==10.1.0
scikit-learn==1.3.2
tensorflow==2.14.0
joblib==1.3.2
gTTS==2.4.0
SpeechRecognition==3.10.0
pydantic==2.4.2